[DEVP-334] split tests into individual file
