+++ shipcat_cli/src/audit.rs
@@ -125,76 +125,3 @@ pub fn audit(ae: &AuditEvent, audcfg: &AuditWebhook) -> Result<()> {
         .chain_err(&mkerr)?;
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    extern crate mockito;
-
-    use std::env;
-    use super::*;
-    use reqwest::Url;
-    use self::mockito::mock;
-    // use self::mockito::Matcher;
-
-    #[test]
-    fn audit_does_ensure_requirements() {
-        // env::set_var("SHIPCAT_AUDIT_CONTEXT_ID", "egcontextid");
-        env::set_var("SHIPCAT_AUDIT_REVISION", "egrevision");
-
-        let audcfg = AuditWebhook{
-            url: Url::parse(&format!("{}/audit", mockito::SERVER_URL)).unwrap(),
-            token: "1234auth".into(),
-        };
-        let wh = Some(Webhooks{
-            audit: audcfg,
-        });
-
-        assert!(ensure_requirements(wh).is_ok());
-    }
-
-    #[test]
-    fn audit_does_audit_deployment() {
-        env::set_var("SHIPCAT_AUDIT_CONTEXT_ID", "egcontextid");
-        env::set_var("SHIPCAT_AUDIT_CONTEXT_LINK", "http://eg.server/");
-        env::set_var("SHIPCAT_AUDIT_REVISION", "egrevision");
-
-        let audcfg = AuditWebhook{
-            url: Url::parse(&format!("{}/audit", mockito::SERVER_URL)).unwrap(),
-            token: "1234auth".into(),
-        };
-        let us = UpgradeState::Completed;
-        let ud = UpgradeData{
-            name: "svc".into(),
-            chart: "wtv".into(),
-            version: "v1".into(),
-            region: "r1".into(),
-            ..Default::default()
-        };
-
-        let mocked = mock("POST", "/audit")
-            .match_header("content-type", "application/json")
-            .match_header("Authorization", "Bearer 1234auth")
-            // TODO: match body with frozen timestamp
-            // .match_body(Matcher::Json(json!(
-            //     {
-            //         "timestamp": "frozen",
-            //         "status": "COMPLETED",
-            //         "context_id": "egcontextid",
-            //         "context_link": "http://eg.server/",
-            //         "type": "deployment",
-            //         "payload": {
-            //             "id": "svc-v1-egrevision",
-            //             "region": "r1",
-            //             "manifestsRevision": "egrevision",
-            //             "service": "svc",
-            //             "version": "v1"
-            //         }
-            //     }
-            // )))
-            .expect(1)
-            .create();
-
-        assert!(audit_deployment(&us, &ud, &audcfg).is_ok());
-        mocked.assert();
-    }
-}
